+++ realm/realm-library/src/main/cpp/util.hpp
@@ -37,7 +37,8 @@
 #include <util/format.hpp>
 
 #include "io_realm_internal_Util.h"
-#include "io_realm_log_LogLevel.h"
+
+#include "jni_util/log.hpp"
 
 #define TRACE               1       // disable for performance
 #define CHECK_PARAMETERS    1       // Check all parameters in API and throw exceptions in java if invalid
@@ -117,53 +118,6 @@ void ThrowNullValueException(JNIEnv* env, realm::Table *table, size_t col_ndx);
 jclass GetClass(JNIEnv* env, const char* classStr);
 
 
-// Debug trace
-extern int trace_level;
-extern jclass realmlog_class;
-extern jmethodID log_trace;
-extern jmethodID log_debug;
-extern jmethodID log_info;
-extern jmethodID log_warn;
-extern jmethodID log_error;
-extern jmethodID log_fatal;
-
-
-// Inspired by From http://www.netmite.com/android/mydroid/system/core/liblog/logd_write.c
-inline void log_message(JNIEnv *env, jmethodID log_method, const char *msg, ...)
-{
-    // Check if a exception has already bee cast. In that case trying to log anything will crash.
-    if (env->ExceptionCheck()) {
-        return;
-    }
-
-    va_list ap;
-    char buf[1024]; // Max logcat line length
-    va_start(ap, msg);
-    // Do formatting in C++. I gave up trying to send C++ variadic arguments back as Java var args.
-    vsnprintf(buf, 1024, msg, ap);
-    va_end(ap);
-
-    jstring log_message = env->NewStringUTF(buf);
-    env->CallStaticVoidMethod(realmlog_class, log_method, log_message, NULL);
-    env->DeleteLocalRef(log_message);
-}
-
-#if TRACE
-    #define TR_ENTER(env) if (trace_level <= io_realm_log_LogLevel_TRACE) { log_message(env, log_trace, " --> %s", __FUNCTION__); } else {}
-    #define TR_ENTER_PTR(env, ptr) if (trace_level <= io_realm_log_LogLevel_TRACE) { log_message(env, log_trace, " --> %s %" PRId64, __FUNCTION__, static_cast<int64_t>(ptr)); } else {}
-    #define TR(env, msg, ...) if (trace_level <= io_realm_log_LogLevel_TRACE) { log_message(env, log_trace, msg, __VA_ARGS__)); } else {}
-    #define TR_ERR(env, msg, ...) if (trace_level <= io_realm_log_LogLevel_ERROR) { log_message(env, log_error, msg, __VA_ARGS__); } else {}
-    #define TR_ERR_NO_VA_ARG(env, msg) if (trace_level <= io_realm_log_LogLevel_ERROR) { log_message(env, log_error, msg); } else {}
-    #define TR_LEAVE(env) if (trace_level <= io_realm_log_LogLevel_TRACE) { log_message(env, log_trace, " <-- %s", __FUNCTION__); } else {}
-#else // TRACE - these macros must be empty
-    #define TR_ENTER(env)
-    #define TR_ENTER_PTR(env, ptr)
-    #define TR(env, msg, ...)
-    #define TR_ERR(env, msg, ...)
-    #define TR_ERR_NO_VA_ARG(env, msg)
-    #define TR_LEAVE(env)
-#endif
-
 // Check parameters
 
 #define TABLE_VALID(env,ptr)    TableIsValid(env, ptr)
@@ -235,7 +189,7 @@ inline bool TableIsValid(JNIEnv* env, T* objPtr)
 
     }
     if (!valid) {
-        TR_ERR(env, "Table %p is no longer attached!", VOID_PTR(objPtr))
+        realm::jni_util::Log::e("Table %1 is no longer attached!", VOID_PTR(objPtr));
         ThrowException(env, IllegalState, "Table is no longer valid to operate on.");
     }
     return valid;
@@ -245,7 +199,7 @@ inline bool RowIsValid(JNIEnv* env, realm::Row* rowPtr)
 {
     bool valid = (rowPtr != NULL && rowPtr->is_attached());
     if (!valid) {
-        TR_ERR(env, "Row %p is no longer attached!", VOID_PTR(rowPtr))
+        realm::jni_util::Log::e("Row %1 is no longer attached!", VOID_PTR(rowPtr));
         ThrowException(env, IllegalState, "Object is no longer valid to operate on. Was it deleted by another thread?");
     }
     return valid;
@@ -259,29 +213,30 @@ bool RowIndexesValid(JNIEnv* env, T* pTable, jlong startIndex, jlong endIndex, j
     if (endIndex == -1)
         endIndex = maxIndex;
     if (startIndex < 0) {
-        TR_ERR(env, "startIndex %" PRId64 " < 0 - invalid!", S64(startIndex))
+        realm::jni_util::Log::e("startIndex %1 < 0 - invalid!", S64(startIndex));
         ThrowException(env, IndexOutOfBounds, "startIndex < 0.");
         return false;
     }
     if (realm::util::int_greater_than(startIndex, maxIndex)) {
-        TR_ERR(env, "startIndex %" PRId64 " > %" PRId64 " - invalid!", S64(startIndex), S64(maxIndex))
+        realm::jni_util::Log::e("startIndex %1 > %2 - invalid!", S64(startIndex), S64(maxIndex));
         ThrowException(env, IndexOutOfBounds, "startIndex > available rows.");
         return false;
     }
 
     if (realm::util::int_greater_than(endIndex, maxIndex)) {
-        TR_ERR(env, "endIndex %" PRId64 " > %" PRId64 " - invalid!", S64(endIndex), S64(maxIndex))
+        realm::jni_util::Log::e("endIndex %1 > %2 - invalid!", S64(endIndex), S64(maxIndex));
         ThrowException(env, IndexOutOfBounds, "endIndex > available rows.");
         return false;
     }
     if (startIndex > endIndex) {
-        TR_ERR(env, "startIndex %" PRId64 " > endIndex %" PRId64 " - invalid!", S64(startIndex), S64(endIndex))
+        realm::jni_util::Log::e(
+                "startIndex %1 > endIndex %2 - invalid!", S64(startIndex), S64(endIndex));
         ThrowException(env, IndexOutOfBounds, "startIndex > endIndex.");
         return false;
     }
 
     if (range != -1 && range < 0) {
-        TR_ERR(env, "range %" PRId64 " < 0 - invalid!", S64(range))
+        realm::jni_util::Log::e("range %1 < 0 - invalid!", S64(range));
         ThrowException(env, IndexOutOfBounds, "range < 0.");
         return false;
     }
@@ -301,7 +256,7 @@ inline bool RowIndexValid(JNIEnv* env, T pTable, jlong rowIndex, bool offset=fal
         size -= 1;
     bool rowErr = realm::util::int_greater_than_or_equal(rowIndex, size);
     if (rowErr) {
-        TR_ERR(env, "rowIndex %" PRId64 " > %" PRId64 " - invalid!", S64(rowIndex), S64(size))
+        realm::jni_util::Log::e("rowIndex %1 > %2 - invalid!", S64(rowIndex), S64(size));
         ThrowException(env, IndexOutOfBounds,
             "rowIndex > available rows: " +
             num_to_string(rowIndex) + " > " + num_to_string(size));
@@ -328,7 +283,8 @@ inline bool ColIndexValid(JNIEnv* env, T* pTable, jlong columnIndex)
     }
     bool colErr = realm::util::int_greater_than_or_equal(columnIndex, pTable->get_column_count());
     if (colErr) {
-        TR_ERR(env, "columnIndex %" PRId64 " > %" PRId64 " - invalid!", S64(columnIndex), S64(pTable->get_column_count()))
+        realm::jni_util::Log::e(
+                "columnIndex %1 > %2 - invalid!", S64(columnIndex), S64(pTable->get_column_count()));
         ThrowException(env, IndexOutOfBounds, "columnIndex > available columns.");
     }
     return !colErr;
@@ -370,7 +326,7 @@ inline bool TblIndexInsertValid(JNIEnv* env, T* pTable, jlong columnIndex, jlong
         return false;
     bool rowErr = realm::util::int_greater_than(rowIndex, pTable->size()+1);
     if (rowErr) {
-        TR_ERR(env, "rowIndex %" PRId64 " > %" PRId64 " - invalid!", S64(rowIndex), S64(pTable->size()))
+        realm::jni_util::Log::e("rowIndex %1 > %2 - invalid!", S64(rowIndex), S64(pTable->size()));
         ThrowException(env, IndexOutOfBounds,
             "rowIndex " + num_to_string(rowIndex) +
             " > available rows " + num_to_string(pTable->size()) + ".");
@@ -384,7 +340,7 @@ inline bool TypeValid(JNIEnv* env, T* pTable, jlong columnIndex, int expectColTy
     size_t col = static_cast<size_t>(columnIndex);
     int colType = pTable->get_column_type(col);
     if (colType != expectColType) {
-        TR_ERR(env, "Expected columnType %d, but got %d.", expectColType, pTable->get_column_type(col))
+        realm::jni_util::Log::e("Expected columnType %1, but got %2.", expectColType, pTable->get_column_type(col));
         ThrowException(env, IllegalArgument, "ColumnType invalid.");
         return false;
     }
@@ -400,7 +356,8 @@ inline bool TypeIsLinkLike(JNIEnv* env, T* pTable, jlong columnIndex)
         return true;
     }
 
-    TR_ERR(env, "Expected columnType %d or %d, but got %d", realm::type_Link, realm::type_LinkList, colType)
+    realm::jni_util::Log::e(
+            "Expected columnType %1 or %2, but got %3", realm::type_Link, realm::type_LinkList, colType);
     ThrowException(env, IllegalArgument, "ColumnType invalid: expected type_Link or type_LinkList");
     return false;
 }
@@ -423,7 +380,7 @@ inline bool ColIsNullable(JNIEnv* env, T* pTable, jlong columnIndex)
         return true;
     }
 
-    TR_ERR_NO_VA_ARG(env, "Expected nullable column type")
+    realm::jni_util::Log::e("Expected nullable column type");
     ThrowException(env, IllegalArgument, "This field is not nullable.");
     return false;
 }